% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tif_to_gif.R
\name{tif_to_gif}
\alias{tif_to_gif}
\title{tif_to_gif: Create an animated GIF file from a folder of TIF files}
\usage{
tif_to_gif(
  prefix,
  suffix,
  frames_per_sec = 1,
  path_to_folder,
  output_location,
  output_name,
  colors = "Purples"
)
}
\arguments{
\item{prefix}{The prefix of your TIF files (argument optional)}

\item{suffix}{The suffix of your TIF files before the file extension (argument optional)}

\item{frames_per_sec}{How fast you want the GIF to go aka frames per second (argument optional)}

\item{path_to_folder}{The location of all TIF files (argument required)}

\item{output_location}{The file path to place the GIF in (argument required)}

\item{output_name}{The name of the output file will be: output_name + _ + prefix (if present) + suffix (if present) + .gif. (argument required)}

\item{colors}{The color palette for your GIF (argument optional)}
}
\value{
A single GIF file made from your TIF files
}
\description{
This function allows the user to generate a GIF from a folder of TIF files. Required inputs are the file paths for the input TIFs and the output GIF, and a name to use as a prefix for the output GIF.
}
\author{
Lauren Harris
}
